refactor(ui): extract backend provider config into reusable components
Each backend provider page's config body (runtime card + auth/catalog +
test panels, with their state/effects) moves into
ui/src/components/settings/providers/{Claude,Codex,Opencode}ProviderConfig.tsx,
plus a BackendProviderConfig dispatcher. The three settings pages slim to
SettingsPageShell + <XxxProviderConfig/>. Pure lift — no behavior change on the
settings routes — so the setup wizard modal can render the same component with
zero copy-paste.
